Refactorings and removal of some old code.
