+++ .idea/misc.xml
@@ -4,8 +4,11 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/activity_main.xml" value="0.2296195652173913" />
+        <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/activity_playlist.xml" value="0.2223731884057971" />
         <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/fragment_all_songs.xml" value="0.2223731884057971" />
         <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/fragment_player.xml" value="0.2296195652173913" />
+        <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/fragment_playlists_tab.xml" value="0.20871327254305977" />
+        <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/layout/list_item_playlist.xml" value="0.20871327254305977" />
         <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/menu/main.xml" value="0.24739583333333334" />
         <entry key="..\:/Users/Ordem/AndroidStudioProjects/SocyMusic/app/src/main/res/xml/preferences.xml" value="0.20153985507246377" />
         <entry key="..\:/Users/tech.assistant2/Desktop/Benji/git/SocyMusic/app/src/main/res/drawable/about_background.xml" value="0.1671875" />
