+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,6 @@
 
         <activity android:name="com.musicplayer.SocyMusic.ui.settings.SettingsActivity"/>
         <activity android:name="com.musicplayer.SocyMusic.ui.dir_browser.DirBrowserActivity"/>
-        <activity android:name="com.musicplayer.SocyMusic.ui.playlist.PlaylistActivity"/>
 
         <service
             android:name="com.musicplayer.SocyMusic.MediaPlayerService"
