+++ app/src/main/java/com/musicplayer/SocyMusic/ui/all_songs/SongHolder.java
@@ -1,4 +1,4 @@
-package com.musicplayer.SocyMusic.ui.allsongs;
+package com.musicplayer.SocyMusic.ui.all_songs;
 
 import android.view.View;
 import android.widget.TextView;
