--- app/src/main/java/com/musicplayer/SocyMusic/ui/playlist/PlaylistFragment.java
@@ -1,18 +1,0 @@
-package com.musicplayer.SocyMusic.ui.playlist;
-
-import android.os.Bundle;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-// Uncompleted.
-// Should display a list of playlists just as MainActivity is displaying a list of songs
-// Refer to Playlist.java, PlaylistData.java and layout/list_item_playlist.xml
-
-public class PlaylistFragment extends Fragment {
-
-    @Override
-    public void onCreate(@Nullable @org.jetbrains.annotations.Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
